# Mixed request analysis