There is no need to filter requests per thread since not all may be hit on the same thread.

Before:
```
λ warp analyze warp-remote-2020-01-28[191406]-unm4.csv.zst -requests
Mixed operations.
[...]
Operation: STAT
 * 106.02 obj/s (59.955s, starting 19:13:04 UTC) (30.0% of operations)

Throughput by host:
 * http://minio1:9000: Avg: 26.15 obj/s (59.948s)
 * http://minio2:9000: Avg: 26.70 obj/s (59.954s)
 * http://minio3:9000: Avg: 26.49 obj/s (59.93s)
 * http://minio4:9000: Avg: 26.69 obj/s (59.94s)

Requests considered: 0:
Not enough requests
 * Avg: 0s 50%: 0s 90%: 0s 99%: 0s Fastest: 0s Slowest: 0s

Requests by host:
 * http://minio1:9000 - 1570 requests:
        - Avg: 117ms Fastest: 14ms Slowest: 550ms 50%: 82ms 90%: 260ms
 * http://minio2:9000 - 1604 requests:
        - Avg: 113ms Fastest: 16ms Slowest: 618ms 50%: 78ms 90%: 251ms
 * http://minio3:9000 - 1591 requests:
        - Avg: 111ms Fastest: 13ms Slowest: 595ms 50%: 79ms 90%: 242ms
 * http://minio4:9000 - 1601 requests:
        - Avg: 117ms Fastest: 14ms Slowest: 630ms 50%: 81ms 90%: 254ms
```

After:
```
λ warp analyze warp-remote-2020-01-28[191406]-unm4.csv.zst -requests
Mixed operations.
[...]
Operation: STAT
 * 106.02 obj/s (59.955s, starting 19:13:04 UTC) (30.0% of operations)

Throughput by host:
 * http://minio1:9000: Avg: 26.15 obj/s (59.948s)
 * http://minio2:9000: Avg: 26.70 obj/s (59.954s)
 * http://minio3:9000: Avg: 26.49 obj/s (59.93s)
 * http://minio4:9000: Avg: 26.69 obj/s (59.94s)

Requests considered: 6352:
 * Avg: 114ms 50%: 80ms 90%: 252ms 99%: 416ms Fastest: 13ms Slowest: 630ms

Requests by host:
 * http://minio1:9000 - 1570 requests:
        - Avg: 117ms Fastest: 14ms Slowest: 550ms 50%: 82ms 90%: 260ms
 * http://minio2:9000 - 1604 requests:
        - Avg: 113ms Fastest: 16ms Slowest: 618ms 50%: 78ms 90%: 251ms
 * http://minio3:9000 - 1591 requests:
        - Avg: 111ms Fastest: 13ms Slowest: 595ms 50%: 79ms 90%: 242ms
 * http://minio4:9000 - 1601 requests:
        - Avg: 117ms Fastest: 14ms Slowest: 630ms 50%: 81ms 90%: 254ms
```

Fixes #3